==> Code/6.AnalyzingCommunities.Rmd <==
---
title: "R Notebook"
output: html_notebook
---
```{r}
library(dplyr)
library(readxl)
library(ggplot2)
```

# First, we read in the datasets of both community assignments and the dataset indexed by community.
```{r}
comms <- read.csv("../Datasets/AttributesByCommunity.csv")
settlements <- read.csv("../datasets/settlements.csv") %>% group_by(UID) %>% summarize(settlementTotal = n(), settlementAmount = sum(settlement))

commAssignments <- read.csv("../Datasets/Officer_Community_Assignments.csv") %>% 
  mutate(Community_ID=as.integer(Community_ID)) %>% 
  left_join(comms %>% select(Community_ID, DetectedCrew, Index_Value)) %>% 
  left_join(settlements, by = "UID")%>%
  mutate(settlementAmount = ifelse(is.na(settlementAmount),0,settlementAmount)) %>% 
  mutate(settlementTotal = ifelse(is.na(settlementTotal),0,settlementTotal)) %>% 
  mutate(DetectedCrew = ifelse(is.na(DetectedCrew),"No",DetectedCrew))
complaints <-read.csv("../Datasets/Complaint_Dataset.csv")

```

# This creates a complainant dataset so we can analyze further trends.
```{r}
blackmales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/blackmales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "Black Male") %>% 
  distinct(CRID, Complainant)

blackfemales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/blackfemales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "Black Female") %>% 
  distinct(CRID, Complainant)

whitemales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/whitemales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "White Male") %>% 
  distinct(CRID, Complainant)

whitefemales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/whitefemales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "White Female") %>% 
  distinct(CRID, Complainant)

hispanicmales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/hispanicmales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "Hispanic Male") %>% 
  distinct(CRID, Complainant)

hispanicfemales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/hispanicfemales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "Hispanic Female") %>% 
  distinct(CRID, Complainant)

naanmales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/naanmales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "NAAN Male") %>% 
  distinct(CRID, Complainant)

naanfemales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/naanfemales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "NAAN Female") %>% 
  distinct(CRID, Complainant)

apimales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/asianpacificislandermales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "API Male") %>% 
  distinct(CRID, Complainant)

apifemales <- read_xlsx("../Datasets/Original Data/ComplainantSubsets/asianpacificislanderfemales.xlsx", sheet = "Allegations") %>% 
  mutate(Complainant = "API Female") %>% 
  distinct(CRID, Complainant)

allComplainants <- blackmales %>% 
  rbind(blackfemales) %>% 
  rbind(whitemales)%>% 
  rbind(whitefemales) %>% 
  rbind(hispanicmales)%>% 
  rbind(hispanicfemales) %>% 
  rbind(naanmales)%>% 
  rbind(naanfemales) %>% 
  rbind(apimales)%>% 
  rbind(apifemales)

#2007 is the first year of good race and gender data

Complainants <- allComplainants %>% count(CRID, Complainant) %>% spread(Complainant, n) %>% mutate(CRID = as.integer(CRID))
Complainants[is.na(Complainants)] <-0
Complainants

crewComplaint <- complaints %>% 
  left_join(commAssignments %>% select(UID, DetectedCrew)) %>% 
  group_by(CRID) %>% 
  summarize(CrewComplaint = ifelse(sum(DetectedCrew=="Yes")>0,"Yes","No")) %>% 
  ungroup()
Complainants2 <- left_join(Complainants, crewComplaint)

Complainants<- allComplainants %>%  mutate(ComplainantID=Complainant) %>% 
  tidyr::separate(ComplainantID, c("Race", "Gender")) %>%
  mutate(CRID=as.integer(CRID)) %>% 
  left_join(crewComplaint)

Complainants2 %>% 
  group_by(CrewComplaint) %>% 
  summarize_all("sum") %>% 
  select(-CRID) %>%
  janitor::clean_names() %>% mutate_at(vars(api_female: white_male), funs(./sum(.)))


complaints %>% left_join(commAssignments) %>% select(CRID, UID, DetectedCrew) %>%  filter(CRID %in% (Complainants2 %>% pull(CRID))) %>% distinct(UID, .keep_all=TRUE) %>% mutate(DetectedCrew = ifelse(is.na(DetectedCrew), "No", DetectedCrew)) %>% count(DetectedCrew) %>% mutate(percent = 100*n/sum(n))

Complainants %>% 
  count(CrewComplaint) %>% 
  mutate(percent = 100*n/sum(n))

Complainants2%>% 
  group_by(CrewComplaint) %>% 
  select(-CRID) %>% 
  summarize_all("sum") %>% 
  select(-CrewComplaint) %>% 
  as.matrix() %>% 
  rowSums()

#Responsible for 24.932% of total Complainants,
16147/(16147+48616)

```

# Crew Officers make up only 3.43% of all officers
```{r}
#3.433% of all officers
nrow(commAssignments %>% filter(DetectedCrew=="Yes"))/nrow(commAssignments)*100
```

# This makes a nice graph that breaks down the complainant distribution for crew officers and non-crew officers
```{r}
temp <- Complainants %>% filter(CrewComplaint == "Yes") %>% count(Complainant)
temp
df <- temp%>% 
  separate(Complainant, c("Race", "Gender")) %>% 
  mutate(percent = 100*n/sum(n)) %>% 
  filter(Race!="API") %>% 
  filter(Race!="NAAN") %>% 
  mutate(percentval=  paste0(format(round(percent, 0), nsmall = 0),"%"))

df %>% ggplot(aes(fill = Gender, label = percentval, x=Race, y= percent))+
  geom_col()+
  geom_text(size = 4,position = position_stack(vjust = 0.5))+
  ggtitle("Complainantization Patterns of Detected Crew Officers\n")+
  xlab("\nComplainant Race")+
  ylab("Percent\n")+
  theme(plot.title = element_text(hjust = 0.5))+
  ylim(c(0,80))


df2 <- Complainants%>%
  count(Complainant) %>% 
  separate(Complainant, c("Race", "Gender")) %>% 
  mutate(percent = 100*n/sum(n)) %>% 
  filter(Race!="API") %>% 
  filter(Race!="NAAN") %>% 
  mutate(percentval=  paste0(format(round(percent,0), nsmall = 0),"%"))

df <- df %>%  mutate(Type = "Crew") 
df2<- df2%>%  mutate(Type = "All") 
dfBoth <- df%>% rbind(df2) %>% select(-n)
dfBoth
dfBoth%>% ggplot(aes(fill = Gender, label = percentval, x=Race, y= percent))+
  geom_col()+
  geom_text(size = 4,position = position_stack(vjust = 0.5))+
  ggtitle("Complainant Breakdown by Officer Type\n")+
  xlab("\nComplainant Race")+
  ylab("Percent\n")+
  theme(plot.title = element_text(hjust = 0.5))+
  ylim(c(0,80))+
  facet_wrap(~Type)
```
# This lets us do some quick t-tests to identify complainant patterns that are different. You can change the complainant name to check a different complainant type. 
```{r}
temp <- Complainants %>% filter(CrewComplaint=="Yes") %>% mutate(Type = "CREW") %>% rbind(Complainants %>% mutate(Type="ALL")) %>% 
  mutate(var = ifelse(Complainant=="Black Male",1, 0))
t.test(temp$var~ temp$Type)
```


# We want to see if officer race changes based on the findings.

```{r}
temp <- commAssignments %>% mutate(var = race=="BLACK")

t.test(temp$var ~temp$DetectedCrew)
```

# We now take a closer look at the settlement data
```{r}
settlements2 <- read.csv("../datasets/settlements.csv")
settlements2
settlements2 %>% 
  left_join(commAssignments %>% select(UID, DetectedCrew)) %>% 
  mutate(DetectedCrew = ifelse(is.na(DetectedCrew), "No", DetectedCrew)) %>% 
  group_by(case_id, settlement) %>% 
  summarize(CrewComplaint = ifelse(sum(DetectedCrew=="Yes")>0,"Yes","No")) %>% 
  group_by(CrewComplaint) %>% 
  summarize(total = n(), settlement = sum(settlement))
```

# Take a look at whether the average settlement amount is different depending on crew vs not crew. 
```{r}
t.test(commAssignments$settlementAmount~commAssignments$DetectedCrew)
```


# This tells us how many complaints are "crew" complaints
```{r}
complaints %>% 
  left_join(commAssignments %>% select(UID, DetectedCrew)) %>% 
  group_by(CRID) %>% 
  summarize(CrewComplaint = ifelse(sum(DetectedCrew=="Yes")>0,"Yes","No")) %>% 
  count(CrewComplaint)

crewComplaint <- complaints %>% 
  left_join(commAssignments %>% select(UID, DetectedCrew)) %>% 
  group_by(CRID) %>% 
  summarize(CrewComplaint = ifelse(sum(DetectedCrew=="Yes")>0,"Yes","No"))
```


# Crews are listed on 23.79% of all use of force complaints
```{r}
complaints %>% filter(Category=="Use Of Force") %>% left_join(crewComplaint) %>% count(CrewComplaint) %>% mutate(percent = n*100/sum(n))
```

# Take a look at the mode weighted degree. 
```{r} 
getmode <- function(v) {
   uniqv <- unique(v)
   uniqv[which.max(tabulate(match(v, uniqv)))]
}
getmode(Degrees$WeightedDegree)
```

# Take a look at arrests for crews vs non-crews.
```{r}
arrests <- read.csv("../Datasets/CB_Officer_UID.csv")

arrests %>% 
  left_join(commAssignments %>% select(UID, DetectedCrew)) %>% 
  distinct(UID, .keep_all=TRUE) %>% 
  mutate(DetectedCrew = ifelse(is.na(DetectedCrew), "No", DetectedCrew)) %>% 
  count(DetectedCrew) %>% 
  mutate(percent = n*100/sum(n))
```

# Crew officers are listed on ~15% of arrests despite comprising only ~5% of officers
```{r}
arrests %>% 
  left_join(commAssignments %>% select(UID, DetectedCrew)) %>%
  mutate(DetectedCrew = ifelse(is.na(DetectedCrew), "No", DetectedCrew)) %>% 
  group_by(CB) %>% 
  summarize(CrewArrest = ifelse(sum(DetectedCrew=="Yes")>0,"Yes","No")) %>% 
  count(CrewArrest) %>% 
  mutate(percent = n*100/sum(n))
```

# Crews are 391,585 arrests for 979 officers
# Non Crews are 2,193,391 arrests for 16,579 officers.
```{r}
391585/979
2193391/16579
```

#Look into specific crews

First up is Crew Number 15 (c15)
```{r}
Degrees <- read.csv("../Datasets/NodeDegrees.csv")
comms %>% filter(Community_ID==15)
uids <- commAssignments %>% filter(Community_ID==15) %>% left_join(Degrees %>% select(-X), by= "UID") %>% pull(UID)
complaints %>% filter(UID %in% uids) %>% distinct(CRID, .keep_all = TRUE) %>%  count(Category) %>% arrange(-n)
complaints %>% filter(UID %in% uids) %>% distinct(CRID, .keep_all = TRUE) %>% filter(unit==3) %>% filter(complete.cases(Latitude, Longitude)) %>% filter(Location!= "Police Building") %>% filter(Category %in% c("Use Of Force", "Illegal Search", "False Arrest"))

complaints %>% filter(UID %in% uids) %>% group_by(CRID) %>% filter(n()>1) %>% ungroup() %>% distinct(CRID, .keep_all = TRUE) %>% count(Location) %>% mutate(percent = n*100/sum(n))
read.csv("../datasets/settlements.csv")  %>% mutate(Crew = ifelse(UID%in%uids,"Yes","No")) %>% group_by(case_id) %>% summarize(Crew = ifelse(sum(Crew=="Yes")>0,"Yes","No"), payoutTotal = sum(settlement)) %>% filter(Crew=="Yes") %>% pull(payoutTotal) %>% sum()

complaints %>% filter(UID %in% uids) %>% group_by(CRID) %>% filter(n()>1) %>% ungroup() %>% distinct(CRID, .keep_all=TRUE) %>% count(Category)
read.csv("../datasets/settlements.csv") %>% filter(UID %in% uids) %>% distinct(case_id, .keep_all = TRUE) %>% pull(settlement) %>% sum()

```

Next we have Crew Number 413 (c413)
```{r}
comms%>% filter(Community_ID==413)
commAssignments %>% filter(Community_ID==413) %>% left_join(Degrees %>% select(-X), by= "UID")
uids <- commAssignments %>% filter(Community_ID==413) %>% left_join(Degrees %>% select(-X), by= "UID") %>% pull(UID)
complaints %>% filter(UID %in% uids) %>% distinct(CRID, .keep_all = TRUE) %>%  count(Category) %>% mutate(percent= n*100/sum(n))
complaints %>% filter(UID %in% uids) %>% group_by(CRID) %>% filter(n()>1) %>% ungroup() %>% distinct(CRID, .keep_all=TRUE) %>% count(Finding)
complaints %>% filter(UID %in% uids) %>% count(unit)
read.csv("../datasets/settlements.csv") %>% filter(UID %in% uids) %>% distinct(case_id, .keep_all = TRUE) %>% pull(settlement) %>% sum()

```

Finally we have Crew Number 424 (c424)
```{r}
comms %>% filter(Community_ID==424)
commAssignments %>% filter(Community_ID==424)%>% left_join(Degrees %>% select(-X), by= "UID")
uids <- commAssignments %>% filter(Community_ID==424)%>% left_join(Degrees %>% select(-X), by= "UID") %>% pull(UID)
complaints %>% filter(UID %in% uids) %>% group_by(CRID) %>% filter(n()>1) %>% ungroup() %>% distinct(CRID, .keep_all=TRUE) %>% count(Finding)
read.csv("../datasets/settlements.csv") %>% filter(UID %in% uids) %>% distinct(case_id, .keep_all = TRUE) %>% pull(settlement) %>% sum()
```

# For our appendix, we make distribution plots of the degrees and weighted degrees in the network.
```{r}
Degrees %>% 
  ggplot(aes(x=Degree))+
  geom_density()+
  theme_bw()+ 
  geom_vline(aes(xintercept=mean(Degree)), color="blue", linetype="dashed", size=.5)+
  ggtitle("Degree Distribution")+
  xlab("Degree")+
  ylab("Density")+
  theme(plot.title = element_text(hjust = 0.5))

Degrees %>% 
  ggplot(aes(x=Weighted.Degree))+
  geom_density()+
  theme_bw()+ 
  geom_vline(aes(xintercept=mean(Weighted.Degree)), color="blue", linetype="dashed", size=.5)+
  ggtitle("Weighted-Degree Distribution")+
  xlab("Weighted-Degree")+
  ylab("Density")+
  theme(plot.title = element_text(hjust = 0.5))
```
